LayoutTests/canvas/philip/tests/2d.path.isPointInPath.outside-expected.txt

< [index] >
2d.path.isPointInPath.outside
isPointInPath() works on paths outside the canvas
References: 2d.path.isPointInPath
Actual output:
Passed
